renames and documentation
