+++ src/mrt_collectors/routeviews.rs
@@ -1,8 +1,8 @@
-//! RouteViews collectors information
+//! RouteViews mrt_collectors information
 //!
 //! API source: <https://api.routeviews.org/collector/?format=json>
 
-use crate::collectors::{MrtCollector, MrtCollectorProject, ToMrtCollector};
+use crate::mrt_collectors::{MrtCollector, MrtCollectorProject, ToMrtCollector};
 use anyhow::Result;
 use chrono::DateTime;
 use serde::Deserialize;
@@ -76,7 +76,7 @@ impl ToMrtCollector for RVCollector {
     }
 }
 
-/// Get RouteViews collectors meta information
+/// Get RouteViews mrt_collectors meta information
 pub fn get_routeviews_collectors() -> Result<Vec<MrtCollector>> {
     let data =
         oneio::read_json_struct::<RvData>("https://api.routeviews.org/collector/?format=json")?;
